Renamed all the files for Components and Updated names + Initial custom loader setup
