--- client/src/data.json
@@ -1,12 +1,0 @@
-{
-    "content": {
-        "media": {
-            "url": "https://twitter.com/woodwingsoft/status/951098126831706112",
-            "properties": {
-                "include-media": true
-            },
-            "provider": "Twitter"
-        },
-    },
-    "identifier": "social-media",
-}
