+++ loaders/custom-loader.js
@@ -1,0 +1,5 @@
+const oldJSONData = require('../Data.json')
+
+module.exports = function(source) {
+    return source;
+}
