+++ src/js/component/Generators.js
