+++ src/js/component/LazyLoading.js
@@ -1,10 +1,8 @@
-import { headLines_KEY} from './variable';
+import { headLines_KEY} from './Variable';
 const headlinesNewsData = async headLinesNews => {
     try {
         const resp = await fetch(headLines_KEY);
         return await resp.json();  
-        console.log("Api Response");  
-        
     } catch (error) {
         console.log(error);
     }
