+++ src/js/component/NewsChannel.js
