+++ src/js/component/Spinner.js
