+++ src/js/component/Variable.js
