+++ webpack.common.js
@@ -1,20 +1,23 @@
 const path = require('path');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const FirstWebpackPlugin = require("./plugins/FirstWebpackPlugin");
-//const json = require('./client/src/data.json');
-
 
 module.exports = {
-    entry: './client/src/js/app.js',
+    entry: './src/js/component/App.js',
     output: {
         filename: 'main.js',
         path: path.resolve(__dirname, 'dist')         
-    },    
+    },   
+    resolveLoader: {
+        modules: ['node_modules', path.resolve(__dirname, 'loaders')]
+    },
     module: {
         rules: [
             {   test: /\.js$/,
                 exclude: /node_modules/,
-                loader: "babel-loader"                    
+                use: [
+                    'custom-loader','babel-loader'
+                ]                       
             },
             {
                 test: /\.css$/,
@@ -39,7 +42,7 @@ module.exports = {
     },
     plugins: [        
         new HtmlWebpackPlugin({
-        template: './client/src/index.html'
+        template: './src/index.html'
         }),
         new FirstWebpackPlugin()
     ]
